feat: add new atoms
