+++ src/app/ui/atoms/inputText/index.ts
@@ -1,0 +1,1 @@
+export { default as InputText } from './inputText';
