+++ src/app/ui/atoms/inputText/inputText.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import { InputProps } from './inputText.types';
+
+const InputText = ({ className, placeholder }: InputProps) => {
+  return (
+    <>
+      <input
+        type="text"
+        className={className}
+        placeholder={placeholder}
+        data-testid="search-input"
+      />
+    </>
+  );
+};
+
+export default InputText;
