+++ src/app/ui/atoms/inputText/inputText.types.ts
@@ -1,0 +1,4 @@
+export interface InputProps {
+  placeholder: string;
+  className: string;
+}
